Fix UB introduced by casting difference of two unsigneds to signed

+++ libyul/backends/evm/OptimizedEVMCodeTransform.cpp
@@ -115,7 +115,7 @@ void OptimizedEVMCodeTransform::operator()(CFG::FunctionCall const& _call)
 		else
 			m_assembly.appendJumpTo(
 				getFunctionLabel(_call.function),
-				static_cast<int>(_call.function.get().returns.size() - _call.function.get().arguments.size()) - (_call.canContinue ? 1 : 0),
+				static_cast<int>(_call.function.get().returns.size()) - static_cast<int>(_call.function.get().arguments.size()) - (_call.canContinue ? 1 : 0),
 				AbstractAssembly::JumpType::IntoFunction
 			);
 		if (useReturnLabel)
